docs(git): add git configuration prompt to readme
Add a new section to the README file with instructions on how to configure
the Git prompt to display Git information (branch and status) in the
terminal. A link to a useful resource with an example is also included.

This change adds value to the project by helping developers improve
their productivity and efficiency in using Git.

The README file was modified in this commit.
Added

    New section to README: "Git"
    Instructions on how to configure Git prompt
    Link to a resource with an example