docs(git): add auto-correct feature to git
This change adds the auto-correct feature to git configuration, which
suggests similar git commands when a non-existing command is executed.
The suggested command will run after a specified delay.

Before, running a non-existing git command resulted in an error message,
which was not helpful. With this change, the user can set a delay for
auto-correct to suggest similar git commands that they may have meant to use.

Example usage is added to the README file.