docs(git): add command to move to git root folder
This commit adds a command to get the root folder based on the .git
folder configuration using the git rev-parse --show-toplevel command.
It also adds a command to move to the root folder. Additionally,
it mentions a function called root that can be added to the .bashrc
file to use this command more easily.