docs(git): add command to pull all subfolders
Added a new command to pull all subfolders in a single command.
This is helpful when working with a lot of repositories and want
to maintain them up to date.

The command runs a loop that checks if a subfolder contains a .git
directory and if so, it runs git pull in that directory.
The loop is executed for all subfolders in the current directory.