docs(bash): add tip extracting col from output
This commit adds a new tip to the Bash section of the README.md file.
The tip shows how to extract a column from a tabular output using
the awk command. It also includes a function that simplifies the
process and makes it more convenient to use. The reference link
to the source of the tip is also added.