docs(curl): add curl with retry
This commit adds a new section to the README describing how to use
CURL with retry. A new code block is added to demonstrate the usage
of this feature with an example. The code block uses the --retry flag
to specify the number of times CURL should retry the request in case of failure.