docs(curl): removing URLs that do not exist
This commit adds an example to the README file showing how to use a
Bash script to remove URLs from a file if they do not return a 200
status code. The script uses the CURL command to check the status code
of each URL and removes any URL that does not return a 200 status code.
This will be helpful for cases where a large number of URLs need to be
checked and updated or removed if they are no longer valid.