docs(fund): finding and deleting a folder
This commit adds an example to the README file showing how to use
the `find` command and the `rm` command to locate and delete a folder.
The `find` command searches for all directories with the name
"FOLDER-NAME" and the `rm` command removes those directories.
This will be helpful for cases where a folder needs to be deleted
from the command line and cannot be deleted using a graphical user
interface.