docs(find): finding and deleting a file
This commit adds an example to the README file showing how to use the
`find` command to locate and delete a file. The `find` command searches
for all files with the name "FILE-NAME" and the `-delete` option removes
those files. This will be helpful for cases where a file needs to be
deleted from the command line and cannot be deleted using a graphical
user interface.