docs(bash): add command to get random file line
Add a new command to the README file that shows how to get a random line
from a file using the shuf command. The command uses the -n option to
specify the number of lines to print, which in this case is set to 1.
The commit message follows the conventional commit format.