docs(bash): command to get only duplicate lines
This commit adds a new command to the README.md file that shows how to
get only duplicate lines using the uniq command. The command can be used
by piping the contents of a file to uniq -D to display only the lines
that appear more than once in the file.