docs(bash): tee command output to screen and file
This commit adds a new tip to the README.md file that demonstrates how
to use the tee command in Bash to output the results of a command to
both the screen and a file simultaneously. The example provided shows
how to use ls -al to list the contents of a directory and then pipe the
output to tee to save the results to a file named blah.txt while also
displaying the output on the screen. This can be useful when you need
to save the results of a command for later use while still being able to
see the output in real time. The tip includes a code snippet and
explanation of how and why to use this approach.