docs(regex): example of changing case with regex
This commit adds an example of how to use regex to change the case of a
string to the README.md file. The example shows how to capitalize the
first letter of each word in a list by using regex capture groups and the \u
modifier. The commit also includes a note explaining that other modifiers,
such as \l, can be used to change the case to lowercase.