docs(docker): command copy files to/from container
This commit adds two new commands to the Docker section of the README.md
file. The commands show how to copy files to/from a Docker container.
The first command copies a file from a container to the host and the
second command copies a file from the host to a container.